\name{varTableDlg}
\alias{varTableDlg}
\title{One-way table of a corpus meta-data variable}
\description{Build a one-way contingency table from a corpus's meta-data variable,
             optionally plotting the result.}
\details{This dialog provides a simple way of computing frequencies from a single meta-data
         variable of a \pkg{tm} corpus. It is merely a wrapper around different steps available from
         the Statistics and Plot menus, but operating on the corpus meta-data instead of the active
         data set.}

\seealso{\code{\link{setCorpusVariables}}, \code{\link{meta}}, \code{\link{table}},
         \code{\link{barchart}}, \code{\link{pie}} }
