\name{specificTermsDlg}
\alias{specificTermsDlg}
\title{List terms specific of a document or level}
\description{List terms most associated (positively or negatively) with each document or each
             of a variable's levels.}
\details{Specifig terms reported here are those with the highest Chi-squared statistic
         for the document or variable level considered. Terms that are negatively associated
         with the document/level are signalled by a negative Chi-squared value: this is just a
         printing convention, as by definition Chi-squared values can only be positive.
}

\seealso{\code{\link{DocumentTermMatrix}}, \code{\link{freqTermsDlg}}, \code{\link{restrictTermsDlg}},
         \code{\link{listTerms}}, \code{\link{termsAssocDlg}} }
